removed unnecessary features, added match caching, shortened match labels
removed div page and greeting page.
added shortened match labels "Q", "S", "F" which replace the old match labels, and also replaced the old system with an enum for match types.
proper method for getting match info has been added but is still unused because it still needs to be fixed before it can be used