getmatches restclient works now & TOKEN is a constant
getMatches() now has replaced getMatchesForEvent() fully, which uses a restclient from spring like how it should have been done in the first place.
this concludes the ai slop code. 🎉🎉🎉

getToken() is now used in config to get the token, and can be accessed from the Config's TOKEN constant, so that all the constants are accessed in the same way (This does mean you will have to restart the application for the token to update).